transposition working and improving performance

+++ generator/printer.go
@@ -41,8 +41,8 @@ func PrintMap(state *rules.BoardState) {
 
 		o.WriteString(fmt.Sprintf("%v %c: %v\n", state.Snakes[numSnake].ID, characters[numSnake], s))
 	}
-	for y := state.Height - 1; y >= 0; y-- {
-		for x := byte(0); x < state.Width; x++ {
+	for y := int32(state.Height) - 1; y >= 0; y-- {
+		for x := int32(0); x < int32(state.Width); x++ {
 			o.WriteRune(board[x][y])
 		}
 		o.WriteString("\n")
